Remove Switch plugin config
This mapping was interfering with the ability to "return" from links
(e.g., in help pages, etc.). Removing the configuration for now. I may
even remove Switch entirely, as I find it sounds more useful than it
actually is in practice (at least with JS), but maybe I just haven't
built the right habits with it.