Add auto-pairs plugin
Trying out as an alternative to delimitMate. Overall I love delimitMate,
but frequently run into a bug where `[<CR>` doesn't indent properly when
inside curly braces. I'm not _positive_ if this is due to delimitMate,
but it's certainly possible. Trying this plugin out as an alternative,
as I've seen a few references to it claiming it's better.